single tof/led instances
